Breaking complex class into traits
